refactor: move things from internals dir to root dir
