+++ .flowconfig
@@ -22,9 +22,9 @@ module.system.node.resolve_dirname=node_modules
 module.system.node.resolve_dirname=app/node_modules
 
 # Map images to WebpackAsset tyoe def
-module.name_mapper.extension='png' -> '<PROJECT_ROOT>/internals/flow/WebpackAsset.js.flow'
-module.name_mapper.extension='jpg' -> '<PROJECT_ROOT>/internals/flow/WebpackAsset.js.flow'
-module.name_mapper.extension='svg' -> '<PROJECT_ROOT>/internals/flow/WebpackAsset.js.flow'
+module.name_mapper.extension='png' -> '<PROJECT_ROOT>/flow/WebpackAsset.js.flow'
+module.name_mapper.extension='jpg' -> '<PROJECT_ROOT>/flow/WebpackAsset.js.flow'
+module.name_mapper.extension='svg' -> '<PROJECT_ROOT>/flow/WebpackAsset.js.flow'
 
 # Map webpack resolved module imports to the app directory
 # Example: components/Foo -> `<project_root>/app/components/Foo`
