--- internals/flow/CSSModule.js.flow
@@ -1,3 +1,0 @@
-// @flow
-
-declare export default { [key: string]: string }
