+++ scripts/fetch-lnd-for-packaging.js
