+++ scripts/lnd-release/canary.sh
